+++ x/ecocredit/client/testsuite/tx.go
@@ -30,9 +30,10 @@ type IntegrationTestSuite struct {
 }
 
 const (
-	validMetadata = "AQ=="
-	classId       = "18AV53K"
-	batchId       = "1b1vrHF"
+	validCreditType = "carbon"
+	validMetadata   = "AQ=="
+	classId         = "18AV53K"
+	batchId         = "1b1vrHF"
 )
 
 var validMetadataBytes = []byte{0x1}
@@ -87,6 +88,7 @@ func (s *IntegrationTestSuite) SetupSuite() {
 				[]string{
 					val.Address.String(),
 					val.Address.String(),
+					validCreditType,
 					validMetadata,
 					fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
 				},
@@ -102,10 +104,11 @@ func (s *IntegrationTestSuite) SetupSuite() {
 
 	// Store the first one in the test suite
 	s.classInfo = &ecocredit.ClassInfo{
-		ClassId:  classId,
-		Designer: val.Address.String(),
-		Issuers:  []string{val.Address.String()},
-		Metadata: validMetadataBytes,
+		ClassId:    classId,
+		Designer:   val.Address.String(),
+		Issuers:    []string{val.Address.String()},
+		CreditType: ecocredit.DefaultParams().CreditTypes[0],
+		Metadata:   validMetadataBytes,
 	}
 
 	startDate, err := client.ParseDate("start date", "2021-01-01")
@@ -199,25 +202,31 @@ func (s *IntegrationTestSuite) TestTxCreateClass() {
 			name:           "missing designer",
 			args:           []string{},
 			expectErr:      true,
-			expectedErrMsg: "Error: accepts 3 arg(s), received 0",
+			expectedErrMsg: "accepts 4 arg(s), received 0",
 		},
 		{
 			name:           "missing issuer",
 			args:           []string{val0.Address.String()},
 			expectErr:      true,
-			expectedErrMsg: "Error: accepts 3 arg(s), received 1",
+			expectedErrMsg: "accepts 4 arg(s), received 1",
+		},
+		{
+			name:           "missing credit type",
+			args:           []string{validCreditType},
+			expectErr:      true,
+			expectedErrMsg: "accepts 4 arg(s), received 1",
 		},
 		{
 			name:           "missing metadata",
 			args:           []string{val0.Address.String(), val0.Address.String()},
 			expectErr:      true,
-			expectedErrMsg: "Error: accepts 3 arg(s), received 2",
+			expectedErrMsg: "accepts 4 arg(s), received 2",
 		},
 		{
 			name:           "too many args",
-			args:           []string{"abcde", "abcde", "abcde", "abcde"},
+			args:           []string{"abcde", "abcde", "abcde", "abcde", "dlskjf"},
 			expectErr:      true,
-			expectedErrMsg: "Error: accepts 3 arg(s), received 4",
+			expectedErrMsg: "accepts 4 arg(s), received 5",
 		},
 		{
 			name: "invalid designer",
@@ -225,6 +234,7 @@ func (s *IntegrationTestSuite) TestTxCreateClass() {
 				[]string{
 					"abcde",
 					val0.Address.String(),
+					validCreditType,
 					validMetadata,
 					makeFlagFrom(val0.Address.String()),
 				},
@@ -239,6 +249,7 @@ func (s *IntegrationTestSuite) TestTxCreateClass() {
 				[]string{
 					val0.Address.String(),
 					"abcde",
+					validCreditType,
 					validMetadata,
 					makeFlagFrom(val0.Address.String()),
 				},
@@ -253,6 +264,7 @@ func (s *IntegrationTestSuite) TestTxCreateClass() {
 				[]string{
 					val0.Address.String(),
 					val0.Address.String(),
+					validCreditType,
 					"=",
 					makeFlagFrom(val0.Address.String()),
 				},
@@ -267,6 +279,7 @@ func (s *IntegrationTestSuite) TestTxCreateClass() {
 				[]string{
 					val0.Address.String(),
 					val0.Address.String(),
+					validCreditType,
 					validMetadata,
 				},
 				s.commonTxFlags()...,
@@ -280,6 +293,7 @@ func (s *IntegrationTestSuite) TestTxCreateClass() {
 				[]string{
 					val0.Address.String(),
 					val0.Address.String(),
+					validCreditType,
 					validMetadata,
 					makeFlagFrom(val0.Address.String()),
 				},
@@ -304,6 +318,7 @@ func (s *IntegrationTestSuite) TestTxCreateClass() {
 						},
 						",",
 					),
+					validCreditType,
 					validMetadata,
 					makeFlagFrom(val0.Address.String()),
 				},