+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * add dates as top level fields in credit batches (#393)
 * add project location as field in credit batches (#394)
 * use dec wrapper for decimal operations (#435)
+* add credit types to credit class (#424)
 * add params for an allowlist of permissioned credit designers (#425)
 
 ## [1.0.0] - 2021-04-13